backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
alembic==1.14.0
python-dotenv==1.0.1
resend
twilio==9.3.7
email-validator==2.1.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
sib-api-v3-sdk==7.6.0

